requirements.txt
flask
flask_cors
requests
pycryptodome
